% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_nifH}
\alias{dataset_nifH}
\title{The dataset_nifH data}
\usage{
data(dataset_nifH)
}
\description{
The dataset_nifH data is a 'microtable' object for the example.
}
\keyword{microtable}
