% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_seqtag_genometag_taxonomy}
\alias{nifH_seqtag_genometag_taxonomy}
\title{The nifH_seqtag_genometag_taxonomy data}
\usage{
data(nifH_seqtag_genometag_taxonomy)
}
\description{
The nifH_seqtag_genometag_taxonomy data is used to store the taxonomic information, sequence tags and genome tags.
}
\keyword{data.frame}
